+++ ListableUI/Sources/Layout/ListLayout/LayoutDescription.swift
@@ -46,9 +46,23 @@ public struct LayoutDescription
     /// Creates a new layout description for the provided layout type, with the provided optional layout configuration.
     public init<LayoutType:ListLayout>(
         layoutType : LayoutType.Type,
-        appearance configure : @escaping (inout LayoutType.LayoutAppearance) -> () = { _ in }
+        appearance configure : (inout LayoutType.LayoutAppearance) -> () = { _ in }
     ) {
-        self.configuration = Configuration(layoutType: layoutType, configure: configure)
+        var layoutAppearance = LayoutType.LayoutAppearance.default
+        configure(&layoutAppearance)
+        
+        self.configuration = Configuration(
+            layoutType: layoutType,
+            layoutAppearance: layoutAppearance
+        )
+    }
+    
+    /// Returns the standard layout properties, which apply to any kind of list layout.
+    ///
+    /// Calling this method is relatively inexpensive – it does not create an instance
+    /// of the backing list layout.
+    public var layoutAppearanceProperties : ListLayoutAppearanceProperties {
+        configuration.layoutAppearanceProperties()
     }
 }
 
@@ -57,7 +71,7 @@ extension ListLayout
 {
     /// Creates a new layout description for a list layout, with the provided optional layout configuration.
     public static func describe(
-        appearance : @escaping (inout Self.LayoutAppearance) -> () = { _ in }
+        appearance : (inout Self.LayoutAppearance) -> () = { _ in }
     ) -> LayoutDescription
     {
         return LayoutDescription(
@@ -74,7 +88,7 @@ extension LayoutDescription
     {
         public let layoutType : LayoutType.Type
         
-        public let configure : (inout LayoutType.LayoutAppearance) -> ()
+        public let layoutAppearance : LayoutType.LayoutAppearance
         
         // MARK: AnyLayoutDescriptionConfiguration
         
@@ -83,10 +97,7 @@ extension LayoutDescription
             behavior: Behavior
         ) -> AnyListLayout
         {
-            var layoutAppearance = LayoutType.LayoutAppearance.default
-            self.configure(&layoutAppearance)
-            
-            return LayoutType(
+            LayoutType(
                 layoutAppearance: layoutAppearance,
                 appearance: appearance,
                 behavior: behavior,
@@ -100,15 +111,16 @@ extension LayoutDescription
             content : (ListLayoutDefaults) -> ListLayoutContent
         ) -> AnyListLayout
         {
-            var layoutAppearance = LayoutType.LayoutAppearance.default
-            self.configure(&layoutAppearance)
-            
-            return LayoutType(
-                layoutAppearance: layoutAppearance,
+            LayoutType(
+                layoutAppearance: self.layoutAppearance,
                 appearance: appearance,
                 behavior: behavior,
                 content: content(LayoutType.defaults)
             )
+        }
+        
+        public func layoutAppearanceProperties() -> ListLayoutAppearanceProperties {
+            .init(self.layoutAppearance)
         }
         
         public func shouldRebuild(layout anyLayout : AnyListLayout) -> Bool
@@ -116,11 +128,7 @@ extension LayoutDescription
             let layout = anyLayout as! LayoutType
             let old = layout.layoutAppearance
             
-            var new = old
-            
-            self.configure(&new)
-            
-            return old != new
+            return old != self.layoutAppearance
         }
         
         public func isSameLayoutType(as anyOther : AnyLayoutDescriptionConfiguration) -> Bool
@@ -149,6 +157,8 @@ public protocol AnyLayoutDescriptionConfiguration
         behavior: Behavior,
         content : (ListLayoutDefaults) -> ListLayoutContent
     ) -> AnyListLayout
+    
+    func layoutAppearanceProperties() -> ListLayoutAppearanceProperties
     
     func shouldRebuild(layout anyLayout : AnyListLayout) -> Bool